# wppconnect-server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "WPPConnect API Rest",
    "description": "Welcome to the wppconnect-server API documentation. This API provides a set of endpoints to interact with the wppconnect-server application, allowing you to build integrations and automate interactions with WhatsApp."
  },
...
  "paths": {
...
    "/api/{session}/start-session": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "startSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhook": {
                    "type": "string"
                  },
                  "waitQrCode": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "webhook": "",
                "waitQrCode": false
              }
            }
          }
        }
      }
    },
...
    "/api/{session}/close-session": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "closeSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/{session}/status-session": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "closeSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}